Run ngrok for five minutes to test that it's actually working
I have a sneaking suspicion that each RUN command in the circle config is self contained. Any processes spawned in a single RUN command will be terminated once that command ends. This should keep the script file alive for five minutes and I can check if ngrok is actually running.